전체 코드 수정 완료
