+++ Mini_TaemProject/src/application/shoulderData.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.text.*?>
 <?import javafx.scene.layout.BorderPane?>
 
-<BorderPane prefHeight="230.0" prefWidth="900.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.ShoulderControl">
+<BorderPane prefHeight="230.0" prefWidth="900.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.ExerciseControl">
    <top>
       <Pane maxHeight="-Infinity" prefHeight="230.0" prefWidth="556.0" style="-fx-background-color: #2A2B40;" BorderPane.alignment="CENTER">
          <children>
